+++ tcp/client/CMakeLists.txt
@@ -4,7 +4,7 @@ project(client)
 set(CMAKE_CXX_STANDARD 14)
 
 set(SOURCE_FILES
-        src/main.cpp ../server/include/request.hpp src/login.cpp src/util.cpp src/main_cycle.cpp include/main_cycle.h)
+        src/main.cpp src/login.cpp src/io_util.cpp src/main_cycle.cpp)
 
 include_directories(../server/include include)
